requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flask==3.0.0
jsonschema==4.17.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==2.0.0
pip==22.3.1
plotly==5.14.1
requests==2.28.1
scipy==1.11.1
urllib3==1.26.13
wheel==0.38.4
pyyaml==6.0 
